+++ huaweicloud-dis-kafka-adapter-common/src/main/java/com/huaweicloud/dis/adapter/kafka/Constants.java
@@ -1,0 +1,36 @@
+/*
+ * Copyright 2002-2010 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.huaweicloud.dis.adapter.kafka;
+
+/**
+ * Created by z00382129 on 2017/11/1.
+ */
+public class Constants {
+
+    public static final String CHECKPOINT_LAST_COMPUTE = "LAST_COMPUTE";
+
+    public static final String CHECKPOINT_LAST_READ = "LAST_READ";
+
+    public static final String CHECKPOINT_COMPUTE_TIMESTAMP = "AT_COMPUTE_TIMESTAMP";
+
+    public static final String CHECKPOINT_LOG_APPEND_TIMESTAMP = "AT_LOG_APPEND_TIMESTAMP";
+    
+    public static final String ERROR_CODE_APP_NAME_NOT_FOUND = "DIS.4332";
+
+    public static final String ERROR_CODE_APP_NAME_EXIST = "DIS.4330";
+    
+}