+++ huaweicloud-dis-kafka-adapter-common/src/main/java/com/huaweicloud/dis/adapter/kafka/model/OffsetResetStrategy.java
@@ -1,0 +1,5 @@
+package com.huaweicloud.dis.adapter.kafka.model;
+
+public enum OffsetResetStrategy {
+    LATEST, EARLIEST, NONE
+}