+++ CHANGELOG.md
@@ -2,4 +2,6 @@
 
 - Features
   * add kafka clients 1.1.0 supports
+  * add kafka clients 0.11 supports
+  * add common adapter